// HumanitarianAssistance.WebApi/OldUI/src/app/model/add-user.ts
export interface AddUsers {
  UserName: any;
  Email: any;
  Password: any;
  FirstName: any;
  LastName: any;
  UserType: any;
  OfficeCode: any;
  OfficeName: any;
  DepartmentId: any;
  DepartmentName: any;
  Status: any;
  Phone: any;
  OfficeId: any;
  UserOfficesList: any;
}

export interface EditUsers {
  UserName: any;
  Email: any;
  // Password : any;
  FirstName: any;
  LastName: any;
  UserType: any;
  OfficeCode: any;
  OfficeName: any;
  DepartmentId: any;
  DepartmentName: any;
  Status: any;
  Id: any;
  Phone: any;
  OfficeId: any;
  // UserOfficesList : any;
}